Add SPA to the set of default processes run for all SCREAMv1 F-Case
This commit adds SPA to the set of processes in the default
scream_input.yaml file.  This will make sure that all F-Compset
runs also include SPA.

NOTE: This PR temporarily hardcodes the default case to be on an ne4
mesh.  SPA requires different mapping files depending on the resolution
of the simulation.  We will need to update the build infrastructure
to accomodate switching files depending on the resolution of the simulation.