Introduce a potential temperature diagnostic and create a unit test.
This commit establishes the "potential temperature" diagnostic.

It also creates the framework for a unit test of that diagnostic.  A
subsequent commit will fill out the unit test to test specific properties.
Currently the test just makes sure that the potential temperature diagnostic
can be constructed, initialized, run and and finalized without throwing an error.