Add infrastructure to pad the SPA source data read from file
By padding the SPA source data in the vertical such that the vertical
data is always 2 indices bigger we are able to force the linear
interpolation routine in EKAT to match the behavior of the linear
interpolation currently used in EAMv2.

The hyam and hybm pressure coordinate system used to construct the
pressure profile for source data has 0 added to beginning and the last
entry is chosen so that the maximum pressure will always be bigger than
the pressure extreme of the target pressure profile.

The data is also padded with 0 in the beginning and a repeat of the last
value at the end.

This will make sure that all target pressures greater than the maximum
source pressure will take the source data surface value.  AT the other
extreme, any pressure that is substantially small will be a percentage
of the top level value of the source data.

See inline comments for more detail.