Update spa_main_test to include explicit test of time and vert interpolation
This commit addresses a reviewer comment to focus the tests in spa_main_test
to explicitly isolate time interpolation and vertical interpolation.

This commit adds a fourth section to the test that does just this.  Additionally,
this commit includes more detailed inline comments on what each section of the
test is examining.