Add an in-house dummy diagnostic to the io.cpp test.
This commit adds a dummy diagnostic to allow us to test for diagnostic
output.  The dummy diagnostic is an arithmetic operation on one of the
fields already tested, and it is added to the INSTANT test case.

Because it is a simple operation on one of the already existent fields
we are able to explicitly test that it was calculated correctly and
written to file correctly.