Add a check for consistency in coupling interval prescribed in MALI and SLM settings
In MALI, coupled interval is setup in namelist.landice under
'config_slm_coupling_interval'. In SLM, it is defined in
namelist.sealevel under the variable 'dtime'. These two variables
should have the same value when running coupled simulations.
This fix checks the consistency between the two values, and gives
an error if the check fails.