Add spatially variable von Mises threshold stress
Add the ability to have the von Mises threshold stress for grounded
and floating ice read in from input files. Add config options
`config_grounded_von_Mises_threshold_stress_source` and
`config_floating_von_Mises_threshold_stress_source` which are set  to
'data' if the spatially varying field is to be red from an input file,
and to 'scalar' if it is to be set by config_grounded_von_Mises_threshold_stress
and config_grounded_von_Mises_threshold_stress.