Create a calving CFL limit and apply to adaptive timestepper
There is reason to think it is incorrect to allow calving to remove more
than a whole grid cell in a given timestep - a CFL condition for
calving.  This commit calculates a calving CFL and adds it to the
adaptive timestepper.  It's important to note that the calving CFL is
calculated from the *previous* timestep, because determining the calving
CFL prior to starting the timestep is prohibitively difficult.  That
would require doing advection and other operations to generate the
geometry that the calvingVelocity will be based on just to calculate the
CFL limit, and then going back and actually progressing through the time
step for real once the dt has been set.  So this addition works under
the principle that the calving CFL limit will not change a lot from
timestep to timestep, thus a lagged CFL limit is a useful way to adapt
the timestep.  In initial tests this appears to work for the simple
MISMIP+ test case.

Note that this additional work is required beyond this commit:
* Right now it is only calculated if using the von Mises calving law
* It is not calculated on the initial time
* Restarts are not handled
* There is not yet an option to disable this
* Adjustments may be desired for how the option for what fraction of the
calving CFL to use interacts with the general CFL dt fraction option
* some general clean up may be desired