This commit adds a temporary fix to the SHOC field checks.
We comment out the check on PBL because for now we are seeing this check
get caught on the first timestep of an ne30 run.  We will want to put this
check back in when we've solved the PBL issue.

We add repairing to the SHOC check for qv.  We are seeing qv go negative
in SHOC.  In order to follow what is done in EAM, which clips negative
water mass, we allow the SHOC check to "repair" qv if it goes out of bounds.