Move mask update prior to calling any calving routines
This commit has no effect on functionality but reduces code redundancy
by calling for mask and geometry update prior to calling any of the
calving routines.  It deletes the mask updates in individual calving
routines as they are no longer needed.  There are no changes to calving
test results with this change, indicating that the mask update is not
necessary, but it is safest to include it to keep the calving module
indpendent of potential changes made elsewhere in the code.