Add config_distribute_unablatedVolumeDynCell option
Add a namelist option to turn distribution of unablatedVolumeDynCell
to neighboring dynamic cells on or off. Default is to have this feature
enabled, but it may be desirable to have it disabled to ensure that results
are not strongly dependent on this ad hoc cleanup, and that the primary
control on the accuracy of calving results comes from the length of
the timestep (i.e., that the code is not violating a CFL-like
condition based on ablation velocity).