Updated to receive pft fractions of actual grid cell
Pft weights must be passed as fraction of actual grid to be
correctly processed by the coupler. These values are normalized
on either side (iac, lnd) to go from/to fraction of land unit/column.
In the land model this requires that higher order weights are
correct before calculation and in some cases renormalization to
unity sums on the land units. Some controls have been updated to
ensure that the components are initialized in the correct order.
domain fractions are now set based on iac-atm interaction. Domain
checking for the iac has been added. Some initialization of iac
data has been moved to coincide with added initialization
procedures. The model runs for a full year, but is failing at the
beginning of year 2.

	modified:   cime/config/e3sm/config_grids.xml
	modified:   cime/src/drivers/mct/cime_config/config_component.xml
	modified:   cime/src/drivers/mct/cime_config/namelist_definition_drv.xml
	modified:   cime/src/drivers/mct/main/cime_comp_mod.F90
	modified:   cime/src/drivers/mct/main/component_mod.F90
	modified:   cime/src/drivers/mct/main/prep_atm_mod.F90
	modified:   cime/src/drivers/mct/main/prep_iac_mod.F90
	modified:   cime/src/drivers/mct/main/seq_domain_mct.F90
	modified:   cime/src/drivers/mct/main/seq_frac_mct.F90
	modified:   cime/src/drivers/mct/shr/seq_infodata_mod.F90
	modified:   cime/src/drivers/mct/shr/seq_timemgr_mod.F90
	modified:   cime/src/share/util/shr_scam_mod.F90
	modified:   components/cam/cime_config/buildnml
	modified:   components/cam/src/cpl/atm_comp_mct.F90
	modified:   components/clm/src/biogeochem/CNEcosystemDynBetrMod.F90
	modified:   components/clm/src/biogeochem/EcosystemDynMod.F90
	modified:   components/clm/src/cpl/lnd_comp_mct.F90
	modified:   components/clm/src/cpl/lnd_import_export.F90
	modified:   components/clm/src/data_types/VegetationType.F90
	modified:   components/clm/src/dyn_subgrid/dynSubgridDriverMod.F90
	modified:   components/clm/src/main/clm_initializeMod.F90
	modified:   components/clm/src/main/controlMod.F90
	modified:   components/clm/src/main/iac2lndMod.F90
	modified:   components/clm/src/main/initGridCellsMod.F90
	modified:   components/clm/src/main/lnd2iacMod.F90
	modified:   components/clm/src/main/subgridRestMod.F90
	modified:   components/gcam/src