Add physics common function to calculate cell length from area.
This commit introduces a new common physics function that will
back out the cell length for a physics column in meters given
the area of the column in radians.

It also introduces a few unit tests for the new function.  The
additional tests also required some rearrangement of the SCALER
set of tests in the common physics unit tests framework.