Adjust namelist variables to match Eulerian transport in dycore.
This commit reverts some of the homme namelist settings to match
the Eulerian dycore, which is what EAMxx is currently set to use.
These can and will be changed in the future when we can use SL
transport.

Additionally, this reverts the SPA data files from a previous commit.
Testing showed that the new SPA data leads to very large nc_activation
which throws a property check error.  Rather than deal with that in
this work we will save investigation of the SPA changes to a subsequent
PR.