Improve logic for calving CFL application
The calving CFL application can be now controlled with the option
config_adaptive_timestep_include_calving.  If that option is true, the
calculation is only applied if a calving law that calculates
calvingVelocity is actually enabled, to avoid attempting to apply the
calvingCFLDt variable when it is not being calculated.

This commit also changes config_adaptive_timestep_CFL_fraction so that
it is only applied to the advective CFL.  That is typically how it has
been used.  I did not change the option name to avoid breaking backwards
compatibility of namelist files.  With this change, there is no longer a
master adjustment - there are only adjustments to the individual
candidate CFL conditions.  Note that the diffusive CFL does not have a
fraction, but it is rarely used, so that is fine for now.