Adjust default SPA data files to reflect the new set of data files.
The old SPA data files had the wrong units for CCN3. which led to a
model with little or no clouds.  This commit adopts updated SPA data
files with the correct units.

This commit also adjusts the bounds for property checks for `nc` in
`P3` and for `nccn_activated` in `SPA` taking into account that with
the correct units these values are expected to be higher.