Apply the new length from area routine to cell length in SHOC
This commit applies the new calculate_dx_from_area routine in the
SHOC interface to convert the area in radians-squared to the proper
length scale in meters.

This commit also addresses reviewer comments about a better variable
name in the new routine and the inclusion of more comments to explain
what the routine is doing.