Switch RossbyRadius GM taper to use a ramp
This merge adds 2 namelist options:
config_GM_Rossby_ramp_min
config_GM_Rossby_ramp_max
that control a ramp used to taper the GM Kappa coefficient based
on the ratio of the cell size (dcEdge) to the Rossby Radius.

The default minimum ratio was chosen so that GM turns off when
the Rossby radius is resolved by 2 grid cells (the previous
threhsold).  The maximum ratio was chosen so that GM is fully on
when a grid cell is 3 Rossby radii in size, corresponding to
~30 km in the Southern Ocean, or a little higher resolution than
the EC30to60 mesh.  This upper limit needs exploration!