Update grounding line mask definition for cells, edges, and vertices
Define grounding line for cells as grounded cells with at least one
dynamic floating neighbor, and for cells and vertices as those with
one grounded and one dynamic floating cell. Previously the grounding
line was defined as where a grounded cell was adjacent to a floating
(dynamic or non-dynamic) or open ocean cell. This definition is
problematic because it prevents us from closing a mass budget, since flux
at the grounding line in the previous definition was not necessarily
leading to loss of grounded ice.