Adjust the scream-scorpio interface freeing pio decomp memory when not needed
This commit adds the new subroutine in the scream_scorpio_interface.F90 code,
`free_decomp`, which will free the memory in scorpio for any pio decomposition
that is no longer needed.  This routine is called whenever a file is closed.

There is also some updates to the infrastructure in the scream scorpio interface
to keep track of which PIO decompositions are still in use, and if one is deleted
having it removed from the list of valid pio decomps.