Fixed use_case files for SSP370 and SSP585
The EAM use_case files for SSP370 and SSP585 contained parameter settings
meant for EAMv1 that need to be cleared. surfdata_map files for ELM are
updated to use one for simyr2015 that is consistent with the starting year
of the SSP scenarios. Several consistency check flags that are typically
used for hybrid runs are specified as default.

The initial SSP370 and SSP585 use_case files went into maint-2.0 via PR #4880.

[BFB] except for SSP370 and SSP585.