Add the do_aerosol_rad runtime option to the radiation interface.
This commit adds the runtime option for the rrtmgp interface to turn
on or off the aerosol forcing to rad, controlled by the boolean parameter
`do_aerosol_rad`.

In the case where `do_aerosol_rad` is false the 4 aerosol forcing fields,
aero_g/ssw/tau_sw and aero_tau_lw are not loaded from the field manager,
instead they are just constructed locally as arrays of zeros.