Fix definition of new parameters in the scream_namelist_defaults
This commit also updates all of the coupled and uncoupled scream unit
tests to take advantage of the new parameters.  In particular, rather
than setting aerosol forcings to 0.0 at initialization via runtime option,
for tests that don't use SPA, use the runtime options these commits
will introduce to avoid using aerosol forcings.

Additionally, this commit cleans up the output parameters for these tests
to just focus on fields that are actually changed during simulation, rather
than those that are set by an initial condition.  The output files are also
formatted to label which output fields are set by which processes.  This will
make it easier to quickly adjust the output streams when doing experiments
where we are adding and removing processes.