EAM stealth feature: "dribbling" for mac-mic subcycles
An option is added to provide tighter coupling between the
cloud macro-microphysics subcycles and the rest of EAM,
as illustrated by Figure 15 in Wan et al. (2021, GMD).
https://gmd.copernicus.org/articles/14/1921/2021/

The tighter coupling reduces time-step sensitivities
in the subtropical marine stratocumulus.

The option is turned on by setting the namelist variable
cld_cpl_opt to an integer larger than 1.

cld_cpl_opt = 1 (default) gives the original scheme in EAMv1 and v2.

[BFB] (unless cld_cpl_opt > 1)