Allocate arrays on all procs before using MPI_GATHER(V)
Intel complained if destination arrays used on MPI_GATHER and
MPI_GATHERV were not allocated before use.  It was not obvious that it
was necessary to do so, because those arguments are unused on all but
the destination processor (where we already have them allocated to the
proper size).  To avoid unnecessary memory usage, they are simply
allocated to length 1 on all other processors.  The deallocate commands
are now performed on all processors.