Fix how precipitation is handled so that it is still correct with subcycling. (#1773)
* Adjust P3 interface to accumulate precip fluxes over subcycles.

This commit adjusts the interface of P3 to accumulate the overall
precipitation flux when P3 is subcycled.  It also resets the accumulated
fluxes to 0.0 whenever a new atmosphere timestep has begun.

Note.  This particular commit is a bit hacky.  There will be a separate PR
coming down the line that will clean up the interface and then we can remove
the hacky changes that this commit includes.

* Switch precip in p3 to a mass, have surface cpl reset it to 0.0

This commit changes the definition of precip_liq_surf and
precip_ice_surf in the P3 interface and in surface coupling to be
a mass with units kg.  As such, the conversion in the surface coupling
export is changed.

Additionally, once the precipitation is used by surface coupling to
export to the component coupler, the surface coupling reset precip
to 0.0 so that the next round of precip accumulation in P3 starts from
0.0.

Note, in order to be able to convert precipitation mass from P3 into a
mass flux this commit added dt information to the surface coupling export.

Similarly, in order to take the precip flux provided by P3 and convert
it to a total mass the dt information needed to be passed to the p3
post-process structure.  For the latter, the dt is set everytime the
p3 run_impl is called.  This will allow for variable timestepping, if
we ever decide to implement that in the code.

* Switch precipitation property check to a lower bound check.

This commit switches the precipitation property checks in P3 to be
lower bound based, since it is difficult to set an upper bound for
total mass accumulated without knowing the timestep.

This commit also adds a comment to the initialization of surface coupling
export and sets the local precipitation variables to be const in post-processing.

* Change the precipitation variable names used locally to be more descriptive.

This commit changes the names of the precipitation variables used in the P3
interface to be more clear.  Local precipitation provided by p3_main is labeled
by `_flux` to denote a flux and the accumulated variable of precipitation that
will be provided to the rest of the model uses `_mass` since it is a mass.

* Append "mass" to the name of precip_liq/ice_surf everywhere in the code.

* forgot to make previous commit change to model_restart_test

* add comment about the switch to lower bound check for precip