Check for dynamic ice that is at the end of non-dynamic peninsula
Do the flood-fill in two steps in remove_icebergs. First, fill from
grounded ice into the dynamic ice shelf. Then, fill from the dynamic
shelf into the non-dynamic cells. This will eliminate the case in
which an iceberg of dynamic cells is connected to the rest of the
ice shelf by non-dynamic cells. This may be an overly expensive
solution to an issue that only happens once in a while, but it
should prevent the velocity solver from failing to converge.