Remove redundant hyam and hybm variables from SPA interface.
There was left over code from an earlier version of SPA which read
the hybrid pressure coordinates from the input file by setting them
as "required" variables.  This was replaced by gathering this data
directly from the SPA data file, but the call to the field manager
was never removed.

This commit fixes this by removing the code.