Fix three diagnostic tests that were failing the cuda_mem_check test on GPU
This commit defines a local variable `num_levs` inside the run_impl for
the dry_static_energy, vertical_layer_interface and vertical_layer_midpoint
diagnostics.  This is then used when calling the common physics functions
that need to know the number of levels in the simulation.

This bug was causing failures in the cuda_mem_check test on GPU machines,
such as weaver.