Add write step tracking to the io unit test so we can test diff frequencies.
This commit updates the io.cpp test to track when a write step should be made, this
is then used to grab the correct input file name.  This is needed for output testing
different output frequencies, since the filename will include the time stamp for the
first timestamp written.  This will be dependent on the output frequency itself.

The choice to track if a write step is going to made allows for flexibility in running
the test.  Allowing us to run for an output frequency that is different than 1 for example.