Update F2010-SCREAMv1 lowres and medres tests
With mpassi for F2010-SCREAMv1, the grids used for the tests
must be replaced because mpassi can not run on atm se grid.
With mosart as river model for the compset and because of its
much larger time stepping, ERS_Ln# and ERP_Ln# tests would always
end up having mosart restart point being ahead of the stop time of
the restart run, which would fail to proceed. Changes to ERS_Ld3
and ERP_Ld3 could work, but it would take longer time. Furthermore
with _Ld3, REST_N=2days is being treated as 2 steps, leading to
much frequent writing of scream.r and mismatched atm restart time stamp
for ERP test's case2run.

To overcome these restart issue with the updated F2010-SCREAMv1, given
river model does not impact the atm results, a new test mod,
atmlndactive-rtm_off, is introduced to turn off the river model. This mode
is used for the updated ERS and ERP tests.

[non-BFB] for all tests using F2010-SCREAMv1 and F2010-SCREAMv1-noAero
          because of the grid changes and the use of the new testmod,
          the updated F2010-SCREAMv1 lowres and medres tests will become
          new tests and the original ones become missing.