Update output frequency control to accept "none" as an option.
This commit is in response to reviewer comments that point out the
default output option used in CIME based runs is "none", not "never".

Although "never" is also a valid option, so this commit expands the
set of output frequencies to use "none" and where applicable updates
logical statements to account for "none" as an option.