Bugfix to re-initialize views in output after each write step
This commit is a bugfix that was discovered when writing output for
averaging type MIN over multiple output snaps.  We were not re-init
the local views so the total minimum remained the same from snap to
snap.

This commit has the output manager re-initialize local views in an
output stream following each write of a timesnap.