First step in using horizontal remapper for data remap in SPA
This commit uses the GSMap information to set the grid for reading
input data.  It also takes the first step towards using the remapper
to actually apply the horizontal remapping.  In this case, the surface
pressure is remapped using the GSMap apply_remap routine.

In order to transition to have the other SPA variables use the horizontal
remapper two things need to happen:
  1. The apply_remap function needs to be extended to support 2+D views.
  2. We need to map the read in data to a similar view that contains
     two extra levels, to accomodate the padding needed for the vertical
     interpolation.