Fix spa test for SP, clean up infrastructure.
This commit is a response to a failed test on weaver for the sp build,
and in response to reviewer comments.

This commit adds more comments for a better understanding of what a
function is used for and how often it is expected to be called.

The fix for the sp fail is to bump up the test tolerance.  The
tolerance is a little high and a better solution would be to revamp
the test to work better for sp.  Currently there it is expected to
use mapping weights following the pattern of 1/2^n where n can be
as large as 20.  At sp this can occur round-off error and cause the
test to false fail.  Bumping up the tolerance is a quick fix to move the
PR along.

There was another set of review comments about changing the name of the
horizontal remapper structure.  That will be added in the next commit.