Switch timeSeriesStats to use simulationStartTime
Previously, timeSeriesStats was using the start time of the current
run as the beginning of stats.  Since this time starts over at
restarts, it was not triggering a reset of the stats as it should
have at the beginning of a restart interval that is commensurate
with the time-averaging interval.

With this merge, we now use the `simulationStartTime` to determine
when time-averaging began, so that alarms are correctly triggered
on restart.