Changes to hydro model to comply with new GL mask change
After various attempts, it was found that the best way to ensure BFB
behavior in the hydro model was to re-create the old GL mask (that
includes both grounded to floating and grounded open ocean) as a new
mask variable in the hydro model.  More work could be done to make a
more univerasl hydro mask and apply it uniformly, but this commit does
the minimum to keep the hydro model BFB using the existing logic in the
hydro model.