Change two instances of cycle to exit
One is new in this branch, the other was longstanding.  In both cases,
the old usage was not hurting anything, but was not actually doing
anything.  With this change, the code will stop looing over neighbors
when it finds what it's looking for, which will be slightly more
efficient.  COMPASS testing indicates no change in results, as should be
the case.