Optionally disable alarm calibration
This merge allows calls to `mpas_set_clock_time` to disable
alarm calibration when setting the current time.  This is useful
because recalibrating alarms has a negative impact on
timeSeriesStats analysis members in MPAS-Ocean and MPAS-Seaice.
The timeSeriesStats AMs get initialized with alarms that should
ring after one compute interval.  If alarms are recalibrated,
they ring immediately instead.  However, the analysis members
typically do not get computed at the start time, so the alarm
actually rings after one time step.  If the time interval desired
for timeSeriesStats is not a single time step, the compute alarm
should not be ringing but it is on the first time step.