Fix an issue where IO couldn't handle the vert interp mask value.
This commit assigns a default mask value for the new diagnostic that is
passed to the vertical interpolation routine.  This value can be changed
by the user if desired.  The reason this is needed is that the default
mask value in vertical interpolation is the numeric limits, which caused IO
to throw an error as a number it can't write to file.

This PR sets a default mask value which is also largely negative, but can be
written to file.  Solves a recently found issue using this diagnostic in CIME
runs where the likelyhood that some columns will be outside of a specific pressure
is more likely.

This commit also adds a test that the diagnostic does produce the masked value
when the pressure level is outside of the bounds.

It also adds a new overload to vertical interpolation allowing the user to specify
a masked value without having to also return an array of masks.