This commit fixes a bug in the team policy declaration for diagnostics.
There are three diagnostics that failed when run in a typical AD case
on a GPU machine.  These three diagnostics;
 - DryStaticEnergy
 - VerticalLayerInterface
 - VerticalLayerMidpoint
all shared one thing in common, under-the-hood they used a Kokkos
parallel scan.  The way the Kokkos policy was set up didn't guarantee
a team size of a power of 2 which caused fails in certain cases.  This
commit solves that by using the appropriate team policy setup.

This commit also solves a separate bug which was discovered where
temporary variables were not being handled properly in the diagnostics
which could lead to non-BFB behavior.