Allow calving laws to operate with restore calving
The primary motivation is to allow mask calving to be applied while
preventing calving margin advance.  But in principle one may also
want to allow calving to cause retreat but not advance.  This is implemented
by not making the calls to li_restore_calving_front and li_calve_ice not be
mutually exclusive and calling restore_calving_front after calving.