Fix indices for remap global id's in coarsening remapper.
This commit fixes an issue where it is assumed that the source grid
has 0-based global id's and that the remap triplets read from file
are 1-based.

Rather than rely on these assumptions, the changes in this commit
calculate the global minimum global id for the source grid and the
triplets and then offset the triplet values to match the source grid.

In addition this commit fixes a few other bugs in the coarsening remapper
that were discovered over the course to development and testing.