Add the capability to produce regional output.
This commit will apply the new coarsening remapper to the scorpio output interface to give the user the ability to produce remapped output on the fly.

In order to trigger regional output the output control YAML file needs to have the "remap_file" parameter given. The value should be a string that points to the remap file to be used.