EAMxx: added member to ResuldAndMsg struct returned by PropertyCheck
The member is a vector of int. If the check fail, and it's a pointwise
field test, the PropertyCheck class can use it to store the indices of
the field entry that failed the check.

This can allow to inspect other fields at the same location where the
check has failed.