Add a way to track alternate field names in output for certain diagnostics.
This commit adds a map to scorpio_output that tracks alternate names that
may exist for a specific output stream.  This is useful for sliced fields which
have a name like X@tom for example but would take the name X_lev_0 under-the-hood.

When using a remapper the io_fm stores the latter name for the field while the field
name itself is queried in `get_field` using the former name.  We didn't catch this before
because the second check would default to calculating the diagnostic as expected.  It
wasn't until we added the check that we only grab diagnostics on the sim_field_mgr that
this issue was found.