This commit updates the vertical remapper to handle N dimension arrays.
The previous implementation was limited to only working with 2D arrays
with layout COL x LEV/ILEV.  This meant we could not easily vertically
remap fields with 3+ dimensions, such as horiz_winds.  Not was it
straightforward to remap a single vertical column.  This set of commits
will address that.

This first commit creates a signature for the main vertical interpolation
routine that supports an Nd view rather than just 2d.  Subsequent commits
will introduce the special 1d view case and will change the signature of
the main routine to be Nd, currently it is still set to 2D.