Update the diagnostics that use the vertical remapper.
This commit fixes the field_at_pressure_level diagnostic to work
with the new signature of the vertically remapped field.

It also comments out the use of the vertically_remapped_field diagnsotic.
It is uncler if we plan to use this diagnostic in the future, so rather
than commiting the time to fix it too we comment it out until we can
decide how to proceed.