Add mask tracking to coarsening remapper.
This commit introduces a new constructor to the coarsening remapper
that allows the user to pass a vector of mask fields which represent
spatial masking in the solution.  The coarsening remapper will then
add the mask fields to the set of remapped fields.

This will make it possible to ignore masked values when remapping. This
will be added in a subsequent commit.

The masked fields are created in the scorpio_output class and passed
to the coarsening remapper.  This commit also fixes a few issues with
the field at horizontal sliced diagnostics that were discovered during
the development of this commit.  They mainly have to do with how the
diagnostic is named and how to query an alternative name, which is
used in the case of the io_fm when horizontally and/or vertically
remapping.