Add a function to rescale masked variables in coarsening remapper.
This commit adds the function rescale_masked_fields in the coarsening
remapper, which is responsible for rescaling and fields whose source
data was masked.  The motivation is that when horizontally remapping
the remap weights should sum to 1.0, but if any of the source data
is masked this effectively means the sum of weights contributing to that
remap is <1.0.

The rescaling operation is straightforward.  In addition to the typical
fields the masks are also remapped.  Since the mask is defined with 0.0
or 1.0 the remapped value will represent the sum of the weights used to
conduct the remap at that column/level.  Rescaling the masked values just
amounts to dividing by the horizontally remapped values.

Note, there is a hard-coded threshold that the masked weight sums has to
be larger than before we rescale.  This avoids dividing by 0.0 or a really
small number.  If a remapped mask is less than the threshold than the field
is assigned a hard-coded "masked value" at that point, which will flag it
as a fully masked output.

This commit rescales only for rank 1 and 2 arrays.  A subsequent commit will
add the rank 3 case to rescaling.