Turn on the ability to conduct vertical and horizontal remap for output.
This commit removes the check that stopped vertical and horizontal remapping
from being used for a single output.

This commit adds infrastructure to scorpio_output, the vertical_remapper and
the coarsening_remapper to facilitate a set of fields that have been both
vertically and horizontally remapped.

Most noteably, vertical remapping can lead to some values being masked.  When
combined with horiztonal remapping this can lead to very strange and unrealistic
values in the final product.  This commit adds the infrastructure to track which
values have been masked in vertical remapping and to remove them from calculation
in the horizontal remapping.

Any horizontal remapping that involves some masked variables is then rescaled so that
the respective weights for those fields that were used will sum to 1.0.  There is also
a check if all of the locations in the horizontal remapping were masked, if so then the
horizontally remapped value is also assigned a mask value.

Finally, this commit adds an IO test to check that vertical+horizontal remapping is
working correctly.