restructure how multiple field managers are handled in scorpio_output
This commit restructures field manager "management" in the scorpio
output class.  before this we had 3 potential field managers, "sim",
"int" and "io", which stood for simulation, intermediate and IO
respectively.

The "int" name was confusing because it is similar to an integer
declaration.  So at the very least this should be changed.  But having
a name like "after vertical remap" was very specific, and would set
the precedent for just adding more and more named field managers if, in
the future, we needed to add even more specialized field managers.

The restructing allows the user to register a field manager pointer
with a unique name which is cataloged for the field manager "getter".
Thus, in the future, if we desire a different specialized field manager
to be added we can do this.