Crusher GPU config: Set default MPIs/GCD to 1
Setting the default MAX_MPI_TASKS_PER_NODE to 8
(corresponding to 1 rank/GCD) as there doesn't
seem to be a sizable benefit from using more ranks/GCD.

From Xingqiu's benchmarking, for 2D gpu runs with ncrms=224,
walltime is 0.314 (8 gcd), 0.30 (16 gcd).
As it stands, we have a 4.67% improvement for using 2 ranks/GCD.

That's almost in the realm of machine variability. Unless we have
more/repeatable experiments that demonstrate a bigger improvement,
setting 1 rank per GCD is reasonable.

We still need to investigate the impact of 8 MPI x 2 OMP vs. 16 MPI
per node. So, leaving MAX_TASKS_PER_NODE to 16 for now.