Drop support for packsize 2^n, n>=1 in vert/horiz remappers.
This commit drops support for an arbitrary packsize = 2^n, n>=1
and just sticks with the two cases of SCREAM_PACK_SIZE and 1. This
should improve performance slightly and makes sense given that we
are not likely to use a packsize not equal to 1 or SCREAM_PACK_SIZE
in the near-future.

This commit also makes some changes to the field_at_pressure_level
diagnositc to remove a reinterpret_cast call that isn't needed.