Add a function to asign attributes to variables in netCDF output.
This commit introduces a new funtion similar to the set_int_attribute
function but this one focuses on adding attributes to variables registered
with a specific netCDF file.

It's primary use, implemented here, is to add a list of subfields as a
variable trait, where applicable.  It can be used for other attributes
that we may want to add on the fly.

This commit also fixes a bug in the set_X_attribute functions that would
throw an error if the `pio_enddef` had not yet been called.