Implement alternating control of Fortran P3 codes for eam and scream
The alternating switch, determined by the case configuration option
(a scream case or else), serves to control which version of Fortran P3
codes being used to build the model. A test is needed to track
scream configuration. The version for the EAM version is currently just
a skelton, which will be replaced by the eventual P3 PR from v3atm

[BFB] A new test added.