Move declaration of output header metadata into C++ code, from F90
This commit takes advantage of the new set_str_attribute_c2f function
to allow the scream_output_manager to handle most of the netCDF
header metadata.  This should make it easier to add information about
the run that should be more readily available to the C++ EAMxx code
and not to the F90 interface.

Note, the institutions metadata is long and would not write properly
using the current definition of set_str_attribute_c2f, so that is
still handled in the F90 code.  This information isn't exapected to
changed often, nor be a runtime option, so this shouldn't be too
much of a headache.

This commit also address a reviewer comment in response to the authors
question.