Add new field to track ice cld frac for output
This commit introduces two new fields to the cloud fraction process which
track the ice and total cloud fraction using a more restrictive threshold
for ice cloud fraction.  These new fields are clearly labeled as "_for_output"
to discern them from the actual cloud fractions to be used by microphysics.

Note, this is needed because the current cloud fraction calculation considers
a cell layer to have 100% ice cloud cover given any amount of cloud ice present.
When considering ice cloud for output variables this has the effect of drowning
out the cloud fractions for any trace amount of cloud ice including falling snow.
So, as issue #2132 illustrates we need to have a separate cloud fraction calculation
with a stricter threshold that can be used for output variables.

Fixes #2132