CLDERA: move call to cldera_compute_stats
The new location *should* be compatible with when EAM
takes a snapshot of its physics state for output.
It's not the same same, but nothing should happen between
the two locations that could alter the physics state,
so for all practical purposes, this new location is the
same as EAM's location for output.