Change description of default algorithm in shr_reprosum_mod module
The comments in the shr_reprosum_mod module describe the default
scalable reproducible algorithm as first converting each floating
point input to a fixed point representation. This is an inadequate
description of the process. (It does adequately describe the original
algorithm in which a single integer was used, but not the current
version which uses a vector of integers representation for each
floating point value.) Comments throughout the module are modified to
increase clarity on this topic.

[BFB] - Bit-For-Bit