Fix logic to prevent overflow in intermediate sums
The default shr_reprosum_mod algorithm, the integer vector algorithm,
first converts each floating point summand into an equivalent
representation using a vector of integers, sums the integer vectors,
then converts the resulting sum back to a floating point
representation. Each integer*8 component IX(i), i=1,... , has an
associated implicit exponent LX(i) and represents the floating point
value real(IX(i),r8)**LX(i). The sequence of exponents all differ by
the same amount, LX(i) - LX(i+1) = max_shift, and are monotonically
decreasing in (i), and each component for a given summand is strictly
less than radix(1_i8)**arr_max_shift in absolute value.

To allow for a larger max_shift (and shorter integer vector), the
components of the integer vector are manipulated after the local
single threaded sums and before the sum over OpenMP threads and MPI
tasks so that
 abs(IX(i)) < radix(1_i8)**max_shift
for all i > 0, as follows. If abs(IX(i)) is larger than or equal to
(radix(1_i8)**arr_max_shift), subtract this 'overlap' from the current
value and add it (appropriately shifted) to IX(i-1). By starting from
the largest index (smallest implicit exponent), this removes any
overlap from each component. Note that the integer vector has an IX(0)
that is initially zero to capture the overlap from IX(1).

The current logic stops the process at index 0. However, this can be
greater than the target radix(1_i8)**max_shift, and the subsequent
summation over MPI tasks and OpenMP threads can potentially overflow
for the distributed sum of IX(0) if the number of OpenMP threads times
the number of MPI tasks is large. By extending the length of the
integer vector from 0:max_levels to -(extra_levels-1):max_levels and
continuing the above process to index -(extra_levels-1), can guarantee
that the integer value for all indices satisfies abs(IX(i)) <
radix(1_i8)**max_shift if choose extra_levels appropriately, thus
eliminating the possibility of overflow.

Another weakness in the original algorithm arises from using

 RX_8 = real(IX(i),r8)
 Overlap = int(scale(RX_8,-max_shift),i8)

to identify the overlap in IX(i), which may miss some of the overlap
if IX(i) has too many significant digits to be represented by a
floating point value. However, simply calculating the overlap using
integer division

 Overlap = IX(i) / (i8_radix**arr_max_shift)

eliminates this (slight) possibility, and also simplifies the code.

This commit implements both fixes. It also augments the comments to
better document why this approach is sufficient to avoid overlap
during the sum over threads and MPI tasks.

Note that, while these changes are not guaranteed to be BFB, the
calculated value for extra_levels for all E3SM simulations in the test
suites (and likely all production simulations) is almost certainly
'1', and so will be identical to the existing algorithm. The change
in the overlap calculation is also likely to be BFB with the current
algorithm in practice for this code fragment and, even if not, is
still unlikely to change the value of the final sum. Again, claiming
that it is BFB for the purposes of this PR so that any non-BFB
behavior in the testing will be examined before the PR is accepted.

BFB