Merge branch 'wlin/atm/p3_alt_eam_scream' (#5496)
The alternating switch, determined by the case configuration option (a scream case or else), serves to control which version of Fortran P3 codes being used to build the model. A test is needed to track scream configuration. The version for the EAM version is currently just a skelton, which will be replaced by the eventual P3 PR from v3atm

[BFB] Two new tests SMS_Ln9_P24x1.ne4_ne4.FDPSCREAM-ARM97
and SMS_Ld1.ne30pg2_ne30pg2.F2010-SCREAM-LR-DYAMOND2