transition all surface exported variables to be helper fields
This commit adds helper fields for all surface coupling exported
variables.  This is in preparation of allowing users to prescribe
surface forcing from the atmosphere.  It's necessary that all variables
be helper fields to avoid changing atmosphere state variables when
prescribing the surface forcing.

This approach is also a bit cleaner since we do not need to have separate
logic to handle fields that are in the field manager for some fields.