Fix surface coupling test, split do_export into different parts.
This commit first fixes issues w/ the surface coupling unit test
that arise from previous commits.

This commit also begins the process of parsing the export fields into
categories of those that will be derived from the EAMxx state and those
that will be set to a constant value.  Subsequent work will focus on
variables that will be updated via a data file.