Update to GCAM6 with functioning land feedback and historical option
Gcam namelist and its defaults have been updated to reflect no hardcoding
of values and the creation of new baselines and their placement in the
inputdata directory.
A compile option has been added to disable gcam parallelization because
the libraries are not present.
Added a check to the harvest array allocation for the iac.
Re-added the atm to iac mapper for the domain.