Change to output in-cloud properties to after the cloud microphysics
This is to incorporate v3atm's PR #25.

The in-cloud properties should be outputted after the cloud microphysics
tendencies are taken into account, whereas they are outputted in P3 without
accounting for the microphysics calculations. This fix does not affect model
simulations, but will affect all of offline analysis where in-cloud
property is used.

[non-BFB]