EAMxx: always write history restart files
Regardless of output frequency and avg type. If avg=INSTANT,
or frequency=1step, or if the last checkpoint coincided with
an output step, then the restart file will not contain any
time record, just some global data that helps the next run
to figure out when the next output will be.