Fix first order flux at ice edge
First order flux needs to be masked when subtracted from higher
order flux, because they don't have the same valid spatial extent.
This seems to largely fix the issue with thickness ramps growing
at the ice margin in tests on the Thwaites domain. It also leads to
a bit more diffusion of the ice mask in the slotted cylinder test,
but it also gets rid of the noisy artifacts in the thickness field,
so that's great!