Updated land feedback to annual average
Now npp, hr, and pft weight are the annual average. This ensures that
the area is aligned with the flux. This brings scalars back into the
zero to 2 range. Also removed some print statements. Also changed the
compile flags for debug to not use strict floating point and to not
error for nans because these occur in GCAM but are not a problem.