Fix ncol/pcols mismatch in linv3_strat_chem_solve
Array elements in fields data structure used by linozv3
has the shape of (pcols,pver). They are used in array operation
for locally declared array of (ncol,pver). ncol and pcols are
not always the same. Run with pm-cpu_gnu debug mode would catch
the mismatch and abort the run. In non-debug mode, or with other
machines/compilers, though the run could proceed, array elements
in the first dimension beyond legitimate ncol would be used (e.g.
shifted to next column or the level below, successively). The
computed results are unintended, if not outright invalid.

Fixed #5661

[NBFB]